refactor(sql-editor): extract the default tab constant
